Yanda Lyu's commit
This commit is to create a constructor method for the Graph class that takes a String parameter filename. This for-loop iterates over each node in the graph and creates a CityNode object for each one. The corresponding element in the this.adjacencyList array is set to the new Edge object.